+++ src/Menu/App/Controllers/MenuItemController.php
@@ -4,6 +4,7 @@ namespace Thinktomorrow\Chief\Menu\App\Controllers;
 
 use Thinktomorrow\Chief\App\Http\Controllers\Controller;
 use Thinktomorrow\Chief\App\Http\Requests\MenuRequest;
+use Thinktomorrow\Chief\Forms\Layouts\PageLayout;
 use Thinktomorrow\Chief\Menu\App\Actions\CreateMenuItem;
 use Thinktomorrow\Chief\Menu\App\Actions\DeleteMenuItem;
 use Thinktomorrow\Chief\Menu\App\Actions\MenuItemApplication;
@@ -12,6 +13,7 @@ use Thinktomorrow\Chief\Menu\App\Queries\MenuTree;
 use Thinktomorrow\Chief\Menu\Menu;
 use Thinktomorrow\Chief\Menu\MenuItem;
 use Thinktomorrow\Chief\Menu\MenuLinkType;
+use Thinktomorrow\Chief\Menu\Resources\MenuItemResource;
 use Thinktomorrow\Chief\Menu\Tree\PrepareMenuItemsForAdminSelect;
 use Thinktomorrow\Chief\Urls\App\Repositories\UrlHelper;
 
@@ -21,7 +23,7 @@ class MenuItemController extends Controller
 
     private MenuItemApplication $menuItemApplication;
 
-    public function __construct(MenuItemApplication $menuItemApplication, PrepareMenuItemsForAdminSelect $prepareMenuItemsForAdminSelect)
+    public function __construct(MenuItemApplication $menuItemApplication, PrepareMenuItemsForAdminSelect $prepareMenuItemsForAdminSelect, private MenuItemResource $resource)
     {
         $this->prepareMenuItemsForAdminSelect = $prepareMenuItemsForAdminSelect;
         $this->menuItemApplication = $menuItemApplication;
@@ -42,12 +44,17 @@ class MenuItemController extends Controller
             MenuTree::byMenu($id),
         );
 
+        $layout = PageLayout::make($this->resource->fields($menuitem))
+            ->model($menuitem)
+            ->filterByNotTagged(['edit', 'not-on-model-create', 'not-on-create']);
+
         return view('chief-menu::create', [
             'menu' => Menu::findOrFail($id),
             'menuitem' => $menuitem,
             'pages' => UrlHelper::allOnlineModels(),
             'ownerReference' => null,
             'parents' => $menuitems,
+            'layout' => $layout,
         ]);
     }
 
@@ -69,7 +76,7 @@ class MenuItemController extends Controller
     {
         $this->authorize('update-page');
 
-        $menuitem = MenuItem::findOrFail($id);
+        $menuitem = MenuItem::with('assetRelation', 'assetRelation.media')->findOrFail($id);
 
         $menuitems = $this->prepareMenuItemsForAdminSelect->prepare(
             MenuTree::byMenu($menuitem->menu_id),
@@ -82,6 +89,7 @@ class MenuItemController extends Controller
             'pages' => UrlHelper::allOnlineModels(),
             'ownerReference' => $menuitem->owner ? $menuitem->owner->modelReference()->getShort() : null,
             'parents' => $menuitems,
+            'layout' => PageLayout::make($this->resource->fields($menuitem))->model($menuitem),
         ]);
     }